+++ python/param_cpt_range.py
@@ -1,4 +1,4 @@
-from numpy import append, sum, array, delete, ndarray, add, vstack, transpose, full, size, mean, var, divide, cumsum, multiply, subtract
+from numpy import append, sum, array, delete, ndarray, add, vstack, transpose, full, size, mean, var, divide, cumsum, multiply, subtract, power
 from sys import exit
 from functions import lapply, which_element
 from param_cpt import seg_len, cpts, cpttype, test_stat
@@ -65,7 +65,7 @@ def sum_function(x):
     if isinstance(x,ndarray) == True:
         p = len(x) #num of cols
         q = len(transpose(x)) # num of rows
-        A = full((q,p),0)
+        A = full((q,p),0, dtype=float)
         for i in range(0,p):
             for j in range(0,q):
                 if x[i][j] > 0:
